% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsl.main.R
\name{der.link}
\alias{der.link}
\title{A subfunction used in estimation}
\usage{
der.link(g.fit, eps = 10^(-4))
}
\arguments{
\item{g.fit}{a \code{mgcv::gam} object}

\item{eps}{a small finite difference used in numerical differentiation.}
}
\description{
This function computes the 1st derivative of the surface-link function with respect to the argument associated with the pure interaction effect term of the smooth, using finite difference.
}
\seealso{
\code{fit.simsl}, \code{simsl}
}
